[make] add hex build, and armclang build support [drivers] add ringbuffer_drv [libc] delete macro FAR, prevent warning [boards] add n32, edit stm32 project to prevent warning [app] add boards_test

+++ apps/gnss_ins_dr_sync/canspeed_obd.h
@@ -14,7 +14,7 @@ void obd_request_speed();
 
 void obd_rx_speed_detect();
 
-uint32_t obd_read_speed();
+int8_t obd_read_speed();
 
 #ifdef __cplusplus
 }
